Cluster profile validation

This function takes multiple random samples from the provided data frame, estimate cluster membership, and calculate the mean (the default, a different statistic can be specified using the summary_fun parameter, e.g. median may be appropriate) for each cluster using both the in sample and out-of-bag (i.e. out of sample). For the out-of-bag sample

Usage

cluster_validation(
  df,
  n_clusters = 2,
  cluster_fun = kmeans,
  get_cluster_fun = function(x) {
x$cluster
 },
  oob_predict_fun = function(fit, newdata) {
     predict(fit, newdata = newdata)
 },
  summary_fun = mean,
  n_samples = 100,
  sample_size = 0.5 * nrow(df),
  replace = FALSE,
  standardize = TRUE,
  seed,
  verbose = interactive(),
  ...
)

# S3 method for class 'clustervalidation'
plot(
  x,
  facet = TRUE,
  plot_complete = TRUE,
  plot_oob = TRUE,
  point_alpha = 0.1,
  point_size = 1,
  line_alpha = 0.1,
  line_width = 1,
  xlab = "",
  ylab = ifelse(attr(x, "standardize"), "Mean Standard Score", "Mean Score"),
  ...
)

# S3 method for class 'clustervalidation'
print(x, ...)

# S3 method for class 'clustervalidation'
summary(object, in_sample = FALSE, oob_sample = TRUE, ...)

# S3 method for class 'cv_summary'
plot(x, ...)

Arguments

df

data frame to estimate clusters. Note that all columns will be used in the estimation.

n_clusters

the number of clusters to estimate.

cluster_fun

the function used to estimate the clusters.

get_cluster_fun

the function used to get the cluster classes. This function takes one parameter, the result of cluster_fun.

oob_predict_fun

the function used to get predictions from the out-of-bag sample. Function takes two parameters, the first is the results of cluster_fun, the second is the out-of-bag sample data frame.

summary_fun

the function used to calculate the statistic for each cluster and iteration. Defaults to mean.

n_samples

the number of random samples to draw.

sample_size

the size of each random sample. Defaults to 50% of observations.

replace

whether sampling should be done with replacement.

standardize

whether the variables should be standardized before estimating clusters.

seed

random number seed. Note that the seed is set before each iteration to seed + i where i is the iteration number.

verbose

whether the function should print the status while running.

...

currently not used.

x

result of summary.clustervalidation()

facet

whether the different samples (i.e. in sample, out-of-bag, and and complete) should be faceted.

plot_complete

whether the profile line using the complete data set should be plotted.

plot_oob

whether the out-of-bag sample should be plotted.

point_alpha

the alpha (transparency) level for points.

line_alpha

the alpha (transparency) level for lines.

xlab

label for the x-axis.

ylab

label for the y-axis.

object

the results of cluster_validation.

in_sample

plot the in sample results.

oob_sample

plot the out-of-bag sample results.

Value

a list with the following elements:

complete_sample

data frame of results using the entire data set.

in_sample

data frame of in sample results.

oob_sample

data frame of out-of-bag results.

complete_model_fit

model fit for the full data set.

in_sample_model_fits

model fits for each sample.

Each of these data frames contain four columns:

iter

the iteration

cluster

the cluster

variable

the variable

value

the mean value of variable for the given cluster and interation

a ggplot2 expression.

a ggplot2 expression.

Details

The number of rows in the resulting data frames will be equal to: ncol(df) * n_samples * n_clusters.